--- src/App.css
@@ -1,6 +1,0 @@
-.App {
-  text-align: center;
-  color: white;
-  min-height: 100vh;
-  background-color: #282c34;
-}
