+++ src/components/Films/Films.jsx
@@ -1,8 +1,10 @@
 import "./films.css"
+import Navbar from "../Navbar/Navbar"
 
 export default function Films() {
   return (
     <div>
+      <Navbar />
       <h1>Films</h1>
     </div>
   )
