+++ src/components/Header/Header.js
@@ -3,7 +3,7 @@ import React from 'react'
 export default function Header(props) {
   return (
     <div className='header-container'>
-      <h1>{props.title}</h1>
+      <h1>{props.appTitle}</h1>
     </div>
   )
 }
