+++ src/components/Main/Main.js
@@ -1,0 +1,12 @@
+import './main.css'
+import Navbar from '../Navbar/Navbar'
+
+
+export default function Main(props) {
+  return (
+    <div>
+    <Navbar />
+      <h1>Homepage</h1>
+    </div>
+  )
+}
