+++ src/components/Main/main.css
@@ -1,0 +1,3 @@
+.main-container {
+  margin-top: 3rem;
+}
