+++ src/components/Navbar/Navbar.js
@@ -7,6 +7,7 @@ export default function Navbar(props) {
       <nav className='navbar'>
         <Link to="/films">Films</Link>
         <Link to="/places">Places</Link>
+        <Link to="/home">Home</Link>
       </nav>
     )
 }
